Format phone number on /review and /confirmation.
